requirements.txt
PyAutoGUI~=0.9.54
Pillow~=9.5.0
PyQt5~=5.15.9
pypdf2~=3.0.1
